requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
passlib[bcrypt]
python-dotenv
langchain-google-genai
langchain-core
pandas
opencv-python
pdfplumber
alembic==1.16.5
coloredlogs==15.0.1
humanfriendly
python-multipart
httpx
watchfiles